fix: made total received size type int64

+++ benchmark/quic/server/main.go
@@ -17,7 +17,7 @@ const ReadTimeout = time.Second * 10
 var (
 	addr    string
 	verbose bool
-	total   int
+	total   int64
 )
 
 func init() {
@@ -66,7 +66,7 @@ func handleSess(sess quic.Session) {
 			return
 		}
 
-		total += nr
+		total += int64(nr)
 		if verbose {
 			fmt.Printf("now %dbyte...\n", total)
 		}
